+++ tests/test_cat2cat_utils.py
@@ -1,11 +1,10 @@
 from cat2cat.datasets import load_trans, load_occup
 from cat2cat import cat2cat
-from cat2cat.dataclass import cat2cat_data, cat2cat_mappings, cat2cat_ml
+from cat2cat.dataclass import cat2cat_data, cat2cat_mappings
 from cat2cat.cat2cat_utils import prune_c2c, dummy_c2c
 
 from pandas import DataFrame
 from numpy import argmax, arange
-from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
 
 trans = load_trans()